Choices can change to another question.
If a choice is created in a question by mistake, we can change it to another
one. But only if this choice hasn't been voted yet.

I had to remove the :counter_cache from :question association because it only
works if you set the Choice's question directly. Not if you set its
question_id. As we'll be changing this through the API, it makes things hard.

I could've added a filter at the controller level (if received a question_id
parameter, load the Question and set it to the changed Choice). But I prefer to
keep this logic into the model.